Stream Config Breakout (#16)
Pull some static methods of the `Picamera2` god-object
